pyproject.toml
[project]
name = "gemini-chat"
version = "0.1.0"
description = "Interactive chat application using Google's Gemini AI model"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",  # For better terminal output
    "typer>=0.9.0",  # For CLI interface
    "pydantic>=2.6.1",  # For data validation
    "pydantic-settings>=2.0.0",  # For settings management
    "structlog>=24.1.0",  # For structured logging
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = false
source-includes = ["tests"]

[tool.pdm.scripts]
start = "python -m gemini_chat.cli chat"
lint = "ruff check ."
format = { composite = ["black .", "isort ."] }
test = "pytest tests/"

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.1",
]

[tool.ruff]
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []
line-length = 100

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
